+++ CHANGELOG.md
@@ -2,13 +2,23 @@
 
 ## --dev/uppcoming--
 
-## 1.0.3-beta - 2025 Sep
+## 1.1.0-beta - 2025 Sep
 ### Parser
 - **Reimplemented parser from scratch** (`core/ASTBuilder.ts`) using the refactored grammar for a much cleaner and more maintainable design.  
 - **Build logic reimplemented** (`core/objectBuilder.ts`) for improved reliability and consistency.  
 - **Error reporting enhanced** to be more user-friendly and informative.  
 - **Fixed bug sometimes wrong line** Sometimes incorrect line number in error messages are now reported accurately.
 - Renamed option `bailSensitivity` to a more shorter and user-friendly name `failLevel`.
+- **Extended `parse` and `parseFile` methods** to support an options-object form.  
+  Example:
+  ```ts
+  const config= YINI.parse(yini, {
+            strictMode: false,
+            failLevel: 'auto',
+            includeMetaData: false,
+            requireDocTerminator: false,
+        })  
+  ```  
 ### Specification Alignment
 - Updated to follow the **latest YINI specification (v1.0.0-rc.3)**.  
   - Document terminator `/END` is now optional in both lenient and strict mode.